feat: FCM 기능 추가

+++ src/main/java/com/example/carebridge/repository/UserAccountRepository.java
@@ -1,6 +1,5 @@
 package com.example.carebridge.repository;
 
-import com.example.carebridge.entity.HospitalInformation;
 import com.example.carebridge.entity.UserAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserAccountRepository extends JpaRepository<UserAccount, Long> {
+public interface UserAccountRepository extends JpaRepository<UserAccount, Integer> {
     Optional<UserAccount> findByPhoneNumber(String phoneNumber);
     UserAccount findByEmail(String email);
 }
